Revert "NOISSUE - Add MainMCU Code"
