--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "mobileplatform/mainmcu/lib/AsynchSerial"]
-	path = mobileplatform/mainmcu/lib/AsynchSerial
-	url = https://github.com/pilotak/AsynchSerial.git
