+++ message.proto
@@ -7,12 +7,13 @@ service ThingsService {
 }
   
 message OmicronMessage {
-	string pitch     = 1; // The pitch for moving back and forth. Forward and backward direction. Int 0 to 2000 i.e 0200. Witch corresponds to -1.00 to +1.00
-	string roll      = 2; // Roll used to move the robot sideways. Left and Right direction. Int 0 to 2000 i.e 1200. Witch corresponds to -1.00 to +1.00
-	string yaw       = 3; // Yaw used to control the direction of the robot. Rotation about perpendicular axis to the floor. Int 0 to 2000 i.e 1000. Witch corresponds to -1.00 to +1.00
-	string publisher = 4; // The publisher of the message. It can be handmotion or mobile. String either h or m
-	string protocol  = 5; // The protocol used to send the message. It can be websocket, http or mqtt. String either w, h or m
+	string pitch     = 1; // The pitch for moving back and forth. Forward and backward direction
+	string roll      = 2; // Roll used to move the robot sideways. Left and Right direction
+	string yaw       = 3; // Yaw used to control the direction of the robot. Rotation about perpendicular axis to the floor
+	string timestamp = 4; // Timestamp used to denote when the message was sent
+	string publisher = 5; // The publisher of the message. It can be handmotion or mobile
+	string protocol  = 6; // The protocol used to send the message. It can be websocket, http or mqtt
 }
-// Message 0200.1200.1000.h.m
+  
 // TODO add respose to robot. To indicate if it is in motion on stationary
 message OmicronResponse {}
