+++ mobileplatform/mainmcu/platformio.ini
@@ -14,15 +14,11 @@ board = genericSTM32F103C8
 framework = mbed
 upload_flags = -c set CPUTAPID 0x2ba01477
 debug_tool = stlink
-debug_server = 
-	$PLATFORMIO_CORE_DIR/packages/tool-openocd/bin/openocd
-	-s $PLATFORMIO_CORE_DIR/packages/tool-openocd/scripts
-	-f interface/stlink.cfg
-	-c "transport select hla_swd"
-	-c "set CPUTAPID 0x2ba01477"
-	-f target/stm32f1x.cfg
-	-c "reset_config none"
-build_flags = 
-	-D PIO_FRAMEWORK_MBED_RTOS_PRESENT
-	-D MBED_CONF_RTOS_PRESENT
-lib_deps = mbed-komaida424/SoftPWM@0.0.0+sha.7918ce37626c
+debug_server =
+  $PLATFORMIO_CORE_DIR/packages/tool-openocd/bin/openocd  
+  -s $PLATFORMIO_CORE_DIR/packages/tool-openocd/scripts
+  -f interface/stlink.cfg
+  -c "transport select hla_swd"
+  -c "set CPUTAPID 0x2ba01477"
+  -f target/stm32f1x.cfg
+  -c "reset_config none"
